requirements.txt
azure-ai-formrecognizer==3.2.0
python-dotenv==1.0.1
